Refactor wizard Form Rendering
